+++ src/Course.php
@@ -2,10 +2,13 @@
 
 namespace Drupal\anu_lms;
 
-use Drupal\Core\Entity\EntityTypeManagerInterface;
+use Drupal\Core\Database\Connection;
 use Drupal\node\NodeInterface;
 use Drupal\Core\Url;
+use Drupal\Core\Entity\EntityRepositoryInterface;
+use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\File\FileUrlGeneratorInterface;
+use Drupal\Core\Entity\EntityStorageInterface;
 
 /**
  * Methods that operate with course nodes.
@@ -17,14 +20,35 @@ class Course {
    *
    * @var \Drupal\Core\Entity\EntityStorageInterface
    */
-  protected $nodeStorage;
+  protected EntityStorageInterface $nodeStorage;
+
+  /**
+   * The paragraph storage.
+   *
+   * @var \Drupal\Core\Entity\EntityStorageInterface
+   */
+  protected EntityStorageInterface $paragraphStorage;
 
   /**
    * The file url generator.
    *
    * @var \Drupal\Core\File\FileUrlGeneratorInterface
    */
-  protected $fileUrlGenerator;
+  protected FileUrlGeneratorInterface $fileUrlGenerator;
+
+  /**
+   * The entity repository service.
+   *
+   * @var \Drupal\Core\Entity\EntityRepositoryInterface
+   */
+  protected EntityRepositoryInterface $entityRepository;
+
+  /**
+   * Database connection.
+   *
+   * @var \Drupal\Core\Database\Connection
+   */
+  protected Connection $database;
 
   /**
    * Constructs service.
@@ -33,10 +57,17 @@ class Course {
    *   The entity type manager.
    * @param \Drupal\Core\File\FileUrlGeneratorInterface $file_url_generator
    *   The file url generator.
+   * @param \Drupal\Core\Entity\EntityRepositoryInterface $entity_repository
+   *   The entity repository.
+   * @param \Drupal\Core\Database\Connection $database
+   *   Database connection object.
    */
-  public function __construct(EntityTypeManagerInterface $entity_type_manager, FileUrlGeneratorInterface $file_url_generator) {
+  public function __construct(EntityTypeManagerInterface $entity_type_manager, FileUrlGeneratorInterface $file_url_generator, EntityRepositoryInterface $entity_repository, Connection $database) {
     $this->nodeStorage = $entity_type_manager->getStorage('node');
+    $this->paragraphStorage = $entity_type_manager->getStorage('paragraph');
     $this->fileUrlGenerator = $file_url_generator;
+    $this->entityRepository = $entity_repository;
+    $this->database = $database;
   }
 
   /**
@@ -45,34 +76,46 @@ class Course {
    * @param \Drupal\node\NodeInterface $course
    *   Course node object.
    *
-   * @return \Drupal\node\NodeInterface|bool
+   * @return \Drupal\node\NodeInterface|null
    *   Lesson or Quiz node object.
    */
-  public function getFirstAccessibleLesson(NodeInterface $course) {
-    $modules = $course->get('field_course_module')->referencedEntities();
+  public function getFirstAccessibleLesson(NodeInterface $course): ?NodeInterface {
+    $modules = $course->get('field_course_module')->getValue();
     foreach ($modules as $module) {
+      $module_paragraph = $this->paragraphStorage->load($module['target_id']);
+
+      // Make sure that the paragraph was successfully loaded.
+      if (empty($module_paragraph)) {
+        continue;
+      }
 
       /** @var \Drupal\node\NodeInterface[] $lessons */
-      $lessons = $module->field_module_lessons->referencedEntities();
+      $lessons = $module_paragraph->get('field_module_lessons')->getValue();
       foreach ($lessons as $lesson) {
-        if ($lesson->access('view')) {
-          return $lesson;
+        /** @var \Drupal\node\NodeInterface $lesson_node */
+        $lesson_node = $this->nodeStorage->load($lesson['target_id']);
+        if ($lesson_node && $lesson_node->access('view')) {
+          return $lesson_node;
         }
       }
 
-      if (!$module->field_module_assessment) {
+      // Ensure that the field with assessment exists, given that it's a part of
+      // submodule anu_lms_assessments.
+      if (!$module_paragraph->hasField('field_module_assessment')) {
         continue;
       }
       /** @var \Drupal\node\NodeInterface[] $quizzes */
-      $quizzes = $module->field_module_assessment->referencedEntities();
+      $quizzes = $module_paragraph->get('field_module_assessment')->getValue();
       foreach ($quizzes as $quiz) {
-        if ($quiz->access('view')) {
-          return $quiz;
+        /** @var \Drupal\node\NodeInterface $quiz_node */
+        $quiz_node = $this->nodeStorage->load($quiz['target_id']);
+        if ($quiz_node && $quiz_node->access('view')) {
+          return $quiz_node;
         }
       }
     }
 
-    return FALSE;
+    return NULL;
   }
 
   /**
@@ -81,34 +124,35 @@ class Course {
    * @param \Drupal\node\NodeInterface $course
    *   Course node object.
    *
-   * @return \Drupal\node\NodeInterface[]
+   * @return array
    *   Flat list of lessons and quizzes.
    */
-  public function getLessonsAndQuizzes(NodeInterface $course) {
+  public function getLessonsAndQuizzes(NodeInterface $course): array {
     $nodes = &drupal_static('anu_lms_course_lessons', []);
     if (!empty($nodes[$course->id()])) {
       return $nodes[$course->id()];
     }
 
+    // Initialize list of lessons and quizzes for the current course.
+    $nodes[$course->id()] = [];
+
     // Get course modules.
+    /** @var \Drupal\paragraphs\ParagraphInterface[] $modules */
     $modules = $course->get('field_course_module')->referencedEntities();
     foreach ($modules as $module) {
-
-      // Get module's lessons.
-      /** @var \Drupal\node\NodeInterface[] $lessons */
-      $lessons = $module->field_module_lessons->referencedEntities();
-      foreach ($lessons as $lesson) {
-        $nodes[$course->id()][] = $lesson;
+      $lesson_values = $module->get('field_module_lessons')->getValue();
+      foreach ($lesson_values as $lesson_value) {
+        $nodes[$course->id()][] = $lesson_value['target_id'];
       }
 
-      // Get module's quiz.
-      if (!$module->field_module_assessment) {
+      // Ensure anu_lms_assessments module is enabled.
+      if (!$module->hasField('field_module_assessment')) {
         continue;
       }
-      /** @var \Drupal\node\NodeInterface[] $quizzes */
-      $quizzes = $module->field_module_assessment->referencedEntities();
-      foreach ($quizzes as $quiz) {
-        $nodes[$course->id()][] = $quiz;
+
+      $quiz_values = $module->get('field_module_assessment')->getValue();
+      foreach ($quiz_values as $quiz_value) {
+        $nodes[$course->id()][] = $quiz_value['target_id'];
       }
     }
 
@@ -137,7 +181,7 @@ class Course {
    * @return \Drupal\Core\Url
    *   Url object for redirect.
    */
-  public function getFinishRedirectUrl(NodeInterface $course) {
+  public function getFinishRedirectUrl(NodeInterface $course): Url {
     $uri = $course->field_course_finish_button->uri;
     if (!empty($uri)) {
       return Url::fromUri($uri);
@@ -151,11 +195,11 @@ class Course {
    * @param \Drupal\node\NodeInterface $course
    *   Course node object.
    *
-   * @return \Drupal\Core\Url
-   *   Url object for redirect.
+   * @return string
+   *   Finish button label.
    */
-  public function getFinishText(NodeInterface $course) {
-    return $course->field_course_finish_button->title;
+  public function getFinishText(NodeInterface $course): string {
+    return $course->field_course_finish_button->title ?? '';
   }
 
   /**
@@ -167,7 +211,7 @@ class Course {
    * @return int
    *   Number of quizzes and lessons.
    */
-  public function countLessons(NodeInterface $course) {
+  public function countLessons(NodeInterface $course): int {
     return count($this->getLessons($course));
   }
 
@@ -180,7 +224,7 @@ class Course {
    * @return int
    *   Number of quizzes.
    */
-  public function countQuizzes(NodeInterface $course) {
+  public function countQuizzes(NodeInterface $course): int {
     return count($this->getQuizzes($course));
   }
 
@@ -193,7 +237,7 @@ class Course {
    * @return \Drupal\node\NodeInterface[]
    *   Lessons in the course.
    */
-  public function getLessons(NodeInterface $course) {
+  public function getLessons(NodeInterface $course): array {
     $modules = $course->get('field_course_module')->referencedEntities();
     $lessonsInCourse = [];
 
@@ -219,7 +263,7 @@ class Course {
    * @return \Drupal\node\NodeInterface[]
    *   Quizzes in the course.
    */
-  public function getQuizzes(NodeInterface $course) {
+  public function getQuizzes(NodeInterface $course): array {
     $modules = $course->get('field_course_module')->referencedEntities();
     $quizzesInCourse = [];
 
@@ -248,7 +292,7 @@ class Course {
    * @return array
    *   List of urls.
    */
-  public function getLessonsAndQuizzesUrls(NodeInterface $course) {
+  public function getLessonsAndQuizzesUrls(NodeInterface $course): array {
     $modules = $course->get('field_course_module')->referencedEntities();
     $urls = [];
 
@@ -257,7 +301,8 @@ class Course {
       $lessons = $module->field_module_lessons->referencedEntities();
       foreach ($lessons as $lesson) {
         if ($lesson->access('view')) {
-          $urls[] = $lesson->toUrl('canonical', ['absolute' => TRUE])->toString();
+          $lesson = $this->entityRepository->getTranslationFromContext($lesson);
+          $urls[] = $lesson->toUrl('canonical')->toString();
         }
       }
 
@@ -268,7 +313,8 @@ class Course {
       $quizzes = $module->field_module_assessment->referencedEntities();
       foreach ($quizzes as $quiz) {
         if ($quiz->access('view')) {
-          $urls[] = $quiz->toUrl('canonical', ['absolute' => TRUE])->toString();
+          $quiz = $this->entityRepository->getTranslationFromContext($quiz);
+          $urls[] = $quiz->toUrl('canonical')->toString();
         }
       }
     }
@@ -280,7 +326,7 @@ class Course {
   }
 
   /**
-   * Returns the accesible audios.
+   * Returns the accessible audios.
    *
    * @param \Drupal\node\NodeInterface $course
    *   Course node object.
@@ -288,24 +334,45 @@ class Course {
    * @return string[]
    *   Audio URLs.
    */
-  public function getAudios(NodeInterface $course) {
-    $lessons = $this->getLessons($course);
-
-    $audiosInCourse = [];
-    foreach ($lessons as $lesson) {
-      $sections = $lesson->get('field_module_lesson_content')->referencedEntities();
-      foreach ($sections as $section) {
-        $sectionContent = $section->get('field_lesson_section_content')->referencedEntities();
-        foreach ($sectionContent as $sectionParagraph) {
-          if ($sectionParagraph->bundle() === 'lesson_audio') {
-            $uri = $sectionParagraph->get('field_audio_file')->entity->getFileUri();
-            $url = $this->fileUrlGenerator->generateAbsoluteString($uri);
-            $audiosInCourse[] = $url;
-          }
-        }
-      }
+  public function getAudios(NodeInterface $course): array {
+    $lesson_ids = [];
+    foreach ($this->getLessons($course) as $lesson) {
+      $lesson_ids[] = $lesson->id();
     }
-    return $audiosInCourse;
+    if (empty($lesson_ids)) {
+      return [];
+    }
+
+    // The select query below relies on Anu LMS data structure to avoid loading
+    // all course paragraph entities.
+    $query = $this->database->select('paragraphs_item_field_data', 'a');
+    // Join the same table to apply conditions to parents of audios.
+    $query->join('paragraphs_item_field_data', 'parent',
+      'a.parent_id = parent.id'
+    );
+    $query
+      ->condition('a.type', 'lesson_audio')
+      ->condition('parent.type', 'lesson_section')
+      ->condition('a.status', 1)
+      ->condition('parent.status', 1)
+      // Keep only audios which parent's parent is one of the course's lessons.
+      ->condition('parent.parent_id', $lesson_ids, 'IN');
+
+    $query->fields('a', ['id']);
+    $query->distinct();
+    $audio_ids = $query->execute()->fetchCol();
+    if (empty($audio_ids)) {
+      return [];
+    }
+
+    $audios_in_course = [];
+    foreach ($this->paragraphStorage->loadMultiple($audio_ids) as $paragraph) {
+      $uri = $paragraph->get('field_audio_file')->entity->getFileUri();
+      $url = $this->fileUrlGenerator->generateAbsoluteString($uri);
+      $audios_in_course[] = $url;
+    }
+
+    return $audios_in_course;
   }
 
 }